Remove pub uses and type (re)definitions, these only make it harder to keep track of things when debugging and refactoring.
