+++ crates/til_query/src/common/logical/logicaltype/group.rs
@@ -127,7 +127,7 @@ impl From<Group> for LogicalType {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::ir::Database;
+    use crate::ir::db::Database;
 
     #[test]
     fn test_new() {
